pyproject.toml
[tool.poetry]
name = "python-flask"
version = "0.1.0"
description = "Start a simple web app with the Flask framework for Python"
authors = ["Google LLC"]
packages = [{include = "src"}]

[tool.poetry.dependencies]
python = "^3.11"
Flask = "^3.0.1"
flask-jwt-extended = "^4.7.1"
python-dotenv = "^1.1.0"
setuptools = "^80.7.1"
flasgger = "^0.9.7.1"
flask-sqlalchemy = "^3.1.1"
pymysql = "^1.1.1"
pydantic = {extras = ["email"], version = "^2.11.4"}
pydantic-settings = "^2.9.1"
flask-migrate = "^4.1.0"

[tool.poetry.scripts]
app = 'src.main:main'

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
